db.BulkExec(): Use ExecContext() instead of Query()
Query() returns a cursor that has to be closed to not block the underlying
database connection. Use ExecContext() instead to both avoid the cursor leak
and also properly pass the context.

This bug lead to the DB connection pool be blocked completely after a certain
number of runtime delete queries.