Allow setting port for PostgreSQL Unix domain socket
When you specify a socket path like /run/postgresql, the actual socket file is
located at /run/postgresql/.s.PGSQL.5432 where the number is the port number of
the server (as it's used for TCP). So it makes sense also allowing to set the
port for Unix domain sockets.

This commit also omits the port from the connection URI if none is given for
TCP connections, thus also using the default value in this case.