utils.IterateOrderedMap for plugin.FormatMessage
The "Tags" and "Extra Tags" printed in the plugin.FormatMessage were
directly read from their map, thus having no order. This resulted in the
same NotificationRequest being represented by different messages due to
the unordered map.

This change was the result of investigating Go's new rangefunc
experiment[0]. The utilization of this novel language feature - which can
also be indirectly used in the absence of `GOEXPERIMENT=rangefunc` - ensures
that the map is traversed in key order.

Closes #177.

[0]: https://go.dev/wiki/RangefuncExperiment