redis: Configurable Database
A Redis server might support multiple databases, allowing to use the
same Redis for different applications. The database can either be
configured for the configuration or later changed via SELECT[0]. By
default, the database 0 will be used.

While both the used go-redis library and even Icinga 2[1] support this
feature, we should expose this here and make this configurable.

As discussed in Icinga DB issue #676[2], this is not something one might
want to encourage all users to use. However, there are definitely setups
where this option might come in handy.

[0]: https://redis.io/docs/latest/commands/select/
[1]: https://github.com/Icinga/icinga2/blob/bca1a8447af26313123183efa863f947f9590731/lib/icingadb/icingadb.ti#L23
[2]: https://github.com/Icinga/icingadb/issues/676