db.DB/redis.Client: Details in Address for GetAddr
The idea of a more informative connection information string was brought
up in #50[0]. Currently, at least Icinga DB and Icinga Notifications are
logging their connection target after startup. This information,
however, is quite vague and was now extended to contain, e.g., if TLS is
used, the username, the database.

> INFO    icingadb        Connecting to database at 'mysql://icingadb@localhost:3306/icingadb'
> INFO    icingadb        Connecting to Redis at 'redis://localhost:6380'

Furthermore, the zapcore.ObjectMarshaler interface was implemented,
allowing both *db.DB and *redis.Client to be used as a zap logging
field, showing the address.

[0] https://github.com/Icinga/icinga-go-library/pull/50#pullrequestreview-2213720430