`logging#Config`: Fix `Options` can no longer be parsed by `go-yaml`
Previously, commit ccf002a introduced `UnmarshalText()` to allow `Options` to
be parsed by the `env` package. With this change, the `go-yaml` package detects
that `Options` should be decoded by that unmarshaller [1, 2]. However, since
`Options` is not a text type, `go-yaml` does not use the unmarshaller [3],
resulting in unmarshalling failures [4]. This change implements
`UnmarshalYAML()` on `Options`, which is tried before `UnmarshalText()` [5],
restoring unmarshalling functionality.

Additionally, it's essential that `Options` is no longer inlined within
`Config`, ensuring that `UnmarshalYAML()` receives only the relevant `options`
section instead of the entire YAML.  Interestingly, this used to work
previously.

[1] https://github.com/goccy/go-yaml/blob/v1.12.0/decode.go#L703
[2] https://github.com/goccy/go-yaml/blob/v1.12.0/decode.go#L833
[3] https://github.com/goccy/go-yaml/blob/v1.12.0/decode.go#L788
[4] https://github.com/goccy/go-yaml/blob/v1.12.0/decode.go#L818
[5] https://github.com/goccy/go-yaml/blob/v1.12.0/decode.go#L763