`config#Load()`: Let error indicate absence of the default config file
If the validation of the configuration fails, `config#Load()` currently only
returns an error message about the failed validation. However, the default
configuration file may be missing, which itself is not an error in the
validation process, as `config#Load()` supports loading configurations from
both YAML and environment variables.  In order to obtain a meaningful error
message, `config#Load()` now returns a combined error message indicating that
the default configuration file is missing, which can be ignored if the
configuration is supposed to be provided entirely via environment variables.