config.TLS: Inline PEM-encoded Certs, Keys and CAs
Introduced support for raw encoded PEM data instead of filenames for
TLS.Cert, TLS.Key and TLS.Ca. This should make it easier to get rid of
the additional entrypoint program for docker-icingadb and allow direct
use of Icinga DB with its environment variable support.

In addition to the implementation, new tests were written with the goal
of complete coverage. A consumer of this type, database.Config, got
another test case using PEMs both as environment variables and as YAML