`redis.Client#XReadUntilResult()`: Respect context errors
Explicitly check for context errors because go-redis v9 does not respect
`context.Canceled` or `context.DeadlineExceeded` unless
`Options.ContextTimeoutEnabled` is set [^1] [^2], which we do not enable.
If the context is canceled or times out during `XRead()` and
there is no data to read, `XRead()` will **still** block until
the block timeout is reached and
return `redis.Nil` instead of the context error.
Without this check, the function would return `redis.Nil`,
potentially leading to unexpected errors for consumers.

With this change, we can also remove the check for no context error when
checking if the error returned by `XRead()` is retryable.

[^1]: https://github.com/redis/go-redis/issues/2556
[^2]: https://github.com/redis/go-redis/issues/2682