logging: ForceLog option to log all messages
Once a zap logger is created, its minimum log level is set in its core.
It is possible to log at this level or at a higher level, for example,
when setting info, one can log at info or error. However, it is not
possible to decrease the level of the logger.

This can be useful when the user has configured a high default level,
but some messages still need to be logged without giving them a
ridiculously high level, e.g., a start log entry with a version string
as an error log entry.