database/redis: More verbosity in OnRetryableError
The OnRetryableError functions came with conditions to only log the
error if it had changed since the last iteration. While this keeps the
log file clean, it hides reoccurring errors and makes it harder to
identify specific errors.

The changed version always logs the reoccurring error and also includes
the duration since the start and the current attempt.