logging: Include error field in journald message
Our journald zapcore.Core logs every zap field as a journald field. This
has the disadvantage that these fields do not shown up in the default
journalctl output, effectively hiding errors. While this is documented,
it is not the expected behavior for most users, resulting in incomplete
bug reports.

This change introduces a new set of zap fields to be also included in
the journald message. Currently, only the "error" field gets this
special treatment, allowing errors to be displayed.

An exemplary Icinga DB log message would change as follows while still
containing the ICINGADB_ERROR field.

> database: Can't connect to database. Retrying
> database: Can't connect to database. Retrying	error="dial tcp [::1]:5432: connect: connection refused"

As a drive-by change, the package-global variable "priorities" was
renamed to "journaldPriorities" to make it more obvious.

Fixes Icinga/icingadb#793.