redis.dialWithLogging: No timeout after connection
Remove the timeout for the Redis connection retry logic after there has
been one successful connection.

Similar to the change in the previous commit to the database connection
logic, the custom Redis dialer keeps track of whether a prior successful
connection has been established. If so, a reconnection after an
interruption during runtime has no timeout.

As a related change, error logging was altered so that only the first
and every tenth error are logged. Unlike the database, the Redis dialer
has a very low backoff interval, resulting in numerous log entries.