Introduce `logging#Error`
The `Error` function returns a `zap.Field` for logging errors. It checks if an
error includes a stack trace from the `pkg/errors` library. If a stack trace is
present, it is suppressed in the log output as logging a stack trace is
unnecessary. Errors without a stack trace are logged normally.