backoff.NewExponentialWithJitter: Ensure bounds
Ensure that the values generated by backoff.NewExponentialWithJitter are
always within [min, max].

While the old function documentation somewhat implied that min would be
an infimum, the smallest possible value would be within [min/2, min) due
to the internal jitter call. Thus, the first Backoff call would always
return a value smaller than min.

Both the documentation and the implementation was slightly changed,
resulting in all durations to be within [min, max].

To ensure this works as intended, a test case was written.