retry: Remove ResetTimeout function
Drop the unused and potentially faulty retry.ResetTimeout function.

The retry.ResetTimeout function relied that the time.Timer's internal
channel is asynchronous. This behavior changed with Go 1.23, resulting
in a now blocking t.C call[^0].

Since there is no usage of this function anymore[^1][^2] and the current
implementation is error prone, it was removed.

> $ for i in icinga{db,-notifications,-kubernetes}; do pushd $i; git pull; git grep ResetTimeout | wc -l; popd; done
> ~/git/github.com/Icinga/icingadb ~/git/github.com/Icinga
> Already up to date.
> 0
> ~/git/github.com/Icinga
> ~/git/github.com/Icinga/icinga-notifications ~/git/github.com/Icinga
> Already up to date.
> 0
> ~/git/github.com/Icinga
> ~/git/github.com/Icinga/icinga-kubernetes ~/git/github.com/Icinga
> Already up to date.
> 0
> ~/git/github.com/Icinga

[^0]: https://pkg.go.dev/time#NewTimer
[^1]: https://github.com/Icinga/icingadb/issues/990
[^2]: https://github.com/Icinga/icingadb/commit/fead08e3e7d5e6277a3aa39878e6de03a5c4907a