logging: allow to provide custom core factory func
It's kinda annoying that the tests spams the GHA logs with lots of log
messages even though the tests aren't failing, see [^1] as an example.
Since zap already provides a logger that logs to the testing.T instance,
allow overriding the core factory to use that logger instead. This way,
the logs are only visible when the test fails or when running the tests
with the `-v` flag.

[^1]: https://github.com/Icinga/icinga-notifications/actions/runs/17466069738/job/49602019404?pr=270