notifications: Simplify and document API
As this is intended to be an user facing API, the API documentation was
a bit extended. Furthermore, some types were simplified.

For example, RulesResp was replaced by a simple string. For an
implementation, this struct did not hold any further required
information not already being part of the map.