notifications: Initial Package Documentation
As the notifications API should be available for external users, a
minimum of API documentation would make a good first impression. This is
a start by adding package level documentation to the notifications
package and each subpackage.

An overview with references to subpackages was written for
notifications, while the user-facing packages got some pointers on how
to get started.

References #62.